+++ perl/LinkedSpec/ActionIR/CanonicalEvents/Core.pm
@@ -145,6 +145,7 @@ my %DIRECT_KIND_BY_CONTRACT_ID = (
  say_stmt                         => 'SAY',
  print_stmt                       => 'PRINT',
  exit_now                         => 'EXIT',
+ next_stmt                        => 'NEXT',
 );
 
 sub _event_args_hash {